+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/).
 Go to the `v2` branch to see the changelog of Lume 2.
 Go to the `v1` branch to see the changelog of Lume 1.
 
+## [3.0.2] - Unreleased
+### Fixed
+- Prism plugin: Allow `prism` as theme name.
+- Add/copy paths starting with `_` (`add("_assets", "assets")`)
+
 ## [3.0.1] - 2025-05-10
 ### Added
 - Reintroduced `site.copy()` function removed in 3.0.0, because it's still useful in specific scenarios.
@@ -123,5 +128,6 @@ Go to the `v1` branch to see the changelog of Lume 1.
 [#749]: https://github.com/lumeland/lume/issues/749
 [#750]: https://github.com/lumeland/lume/issues/750
 
+[3.0.2]: https://github.com/lumeland/lume/compare/v3.0.1...HEAD
 [3.0.1]: https://github.com/lumeland/lume/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/lumeland/lume/releases/tag/v3.0.0